make db orm

+++ src/main/java/pl/pzsp2back/dto/SignUpDto.java
@@ -1,3 +1,4 @@
 package pl.pzsp2back.dto;
 
-public record SignUpDto(String login, String password, String email, String name, String surname, int groupId) {}
+
+public record SignUpDto(String login, String password, String email, String name, String surname, long groupId) {}
